+++ lib/graphql/schema.rb
@@ -82,6 +82,12 @@ module GraphQL
     include GraphQL::Define::InstanceDefinable
     extend GraphQL::Schema::FindInheritedValue
 
+    class DuplicateTypeNamesError < GraphQL::Error
+      def initialize(type_name:, first_definition:, second_definition:, path:)
+        super("Multiple definitions for `#{type_name}`. Previously found #{first_definition.inspect} (#{first_definition.class}), then found #{second_definition.inspect} (#{second_definition.class}) at #{path.join(".")}")
+      end
+    end
+
     class UnresolvedLateBoundTypeError < GraphQL::Error
       attr_reader :type
       def initialize(type:)
@@ -715,14 +721,20 @@ module GraphQL
     # @param using [Hash] Plugins to attach to the created schema with `use(key, value)`
     # @param interpreter [Boolean] If false, the legacy {Execution::Execute} runtime will be used
     # @return [Class] the schema described by `document`
-    def self.from_definition(definition_or_path, default_resolve: BuildFromDefinition::DefaultResolve, interpreter: true, parser: BuildFromDefinition::DefaultParser, using: {})
+    def self.from_definition(definition_or_path, default_resolve: nil, interpreter: true, parser: BuildFromDefinition::DefaultParser, using: {})
       # If the file ends in `.graphql`, treat it like a filepath
       definition = if definition_or_path.end_with?(".graphql")
         File.read(definition_or_path)
       else
         definition_or_path
       end
-      GraphQL::Schema::BuildFromDefinition.from_definition(definition, default_resolve: default_resolve, parser: parser, using: using, interpreter: interpreter)
+      GraphQL::Schema::BuildFromDefinition.from_definition(
+        definition,
+        default_resolve: default_resolve,
+        parser: parser,
+        using: using,
+        interpreter: interpreter,
+      )
     end
 
     # Error that is raised when [#Schema#from_definition] is passed an invalid schema definition string.
@@ -896,6 +908,7 @@ module GraphQL
         schema_defn.cursor_encoder = cursor_encoder
         schema_defn.tracers.concat(tracers)
         schema_defn.query_analyzers.concat(query_analyzers)
+        schema_defn.analysis_engine = analysis_engine
 
         schema_defn.middleware.concat(all_middleware)
         schema_defn.multiplex_analyzers.concat(multiplex_analyzers)
@@ -922,6 +935,11 @@ module GraphQL
         if !schema_defn.interpreter?
           schema_defn.instrumenters[:query] << GraphQL::Schema::Member::Instrumentation
         end
+
+        if new_connections?
+          schema_defn.connections = self.connections
+        end
+
         schema_defn.send(:rebuild_artifacts)
 
         schema_defn
@@ -1703,7 +1721,7 @@ module GraphQL
         end
         late_types = []
         new_types = Array(t)
-        new_types.each { |t| add_type(t, owner: nil, late_types: late_types) }
+        new_types.each { |t| add_type(t, owner: nil, late_types: late_types, path: [t.graphql_name]) }
         missed_late_types = 0
         while (late_type_vals = late_types.shift)
           type_owner, lt = late_type_vals
@@ -1712,13 +1730,13 @@ module GraphQL
             # Reset the counter, since we might succeed next go-round
             missed_late_types = 0
             update_type_owner(type_owner, type)
-            add_type(type, owner: type_owner, late_types: late_types)
+            add_type(type, owner: type_owner, late_types: late_types, path: [type.graphql_name])
           elsif lt.is_a?(LateBoundType)
             if (type = get_type(lt.graphql_name))
               # Reset the counter, since we might succeed next go-round
               missed_late_types = 0
               update_type_owner(type_owner, type)
-              add_type(type, owner: type_owner, late_types: late_types)
+              add_type(type, owner: type_owner, late_types: late_types, path: [type.graphql_name])
             else
               missed_late_types += 1
               # Add it back to the list, maybe we'll be able to resolve it later.
@@ -1793,7 +1811,7 @@ module GraphQL
         end
       end
 
-      def add_type(type, owner:, late_types:)
+      def add_type(type, owner:, late_types:, path:)
         if type.respond_to?(:metadata) && type.metadata.is_a?(Hash)
           type_class = type.metadata[:type_class]
           if type_class.nil?
@@ -1813,46 +1831,52 @@ module GraphQL
 
         if (prev_type = own_types[type.graphql_name])
           if prev_type != type
-            raise ArgumentError, "Conflicting type definitions for `#{type.graphql_name}`: #{prev_type} (#{prev_type.class}), #{type} #{type.class}"
+            raise DuplicateTypeNamesError.new(
+              type_name: type.graphql_name,
+              first_definition: prev_type,
+              second_definition: type,
+              path: path,
+            )
           else
             # This type was already added
           end
         elsif type.is_a?(Class) && type < GraphQL::Schema::Directive
-          type.arguments.each do |_name, arg|
+          type.arguments.each do |name, arg|
             arg_type = arg.type.unwrap
             references_to(arg_type, from: arg)
-            add_type(arg_type, owner: arg, late_types: late_types)
+            add_type(arg_type, owner: arg, late_types: late_types, path: path + [name])
           end
         else
           own_types[type.graphql_name] = type
           if type.kind.fields?
-            type.fields.each do |_name, field|
+            type.fields.each do |name, field|
               field_type = field.type.unwrap
               references_to(field_type, from: field)
-              add_type(field_type, owner: field, late_types: late_types)
-              field.arguments.each do |_name, arg|
+              field_path = path + [name]
+              add_type(field_type, owner: field, late_types: late_types, path: field_path)
+              field.arguments.each do |arg_name, arg|
                 arg_type = arg.type.unwrap
                 references_to(arg_type, from: arg)
-                add_type(arg_type, owner: arg, late_types: late_types)
+                add_type(arg_type, owner: arg, late_types: late_types, path: field_path + [arg_name])
               end
             end
           end
           if type.kind.input_object?
-            type.arguments.each do |_name, arg|
+            type.arguments.each do |arg_name, arg|
               arg_type = arg.type.unwrap
               references_to(arg_type, from: arg)
-              add_type(arg_type, owner: arg, late_types: late_types)
+              add_type(arg_type, owner: arg, late_types: late_types, path: path + [arg_name])
             end
           end
           if type.kind.union?
             own_possible_types[type.graphql_name] = type.possible_types
             type.possible_types.each do |t|
-              add_type(t, owner: type, late_types: late_types)
+              add_type(t, owner: type, late_types: late_types, path: path + ["possible_types"])
             end
           end
           if type.kind.interface?
             type.orphan_types.each do |t|
-              add_type(t, owner: type, late_types: late_types)
+              add_type(t, owner: type, late_types: late_types, path: path + ["orphan_types"])
             end
           end
           if type.kind.object?
@@ -1860,7 +1884,7 @@ module GraphQL
             type.interfaces.each do |i|
               implementers = own_possible_types[i.graphql_name] ||= []
               implementers << type
-              add_type(i, owner: type, late_types: late_types)
+              add_type(i, owner: type, late_types: late_types, path: path + ["implements"])
             end
           end
         end