+++ lib/graphql/static_validation/literal_validator.rb
@@ -6,59 +6,73 @@ module GraphQL
       def initialize(context:)
         @context = context
         @warden = context.warden
+        @invalid_response = GraphQL::Query::InputValidationResult.new(valid: false, problems: [])
+        @valid_response = GraphQL::Query::InputValidationResult.new(valid: true, problems: [])
       end
 
       def validate(ast_value, type)
-        if type.nil?
-          # this means we're an undefined argument, see #present_input_field_values_are_valid
-          maybe_raise_if_invalid(ast_value) do
-            false
-          end
-        elsif ast_value.is_a?(GraphQL::Language::Nodes::NullValue)
-          maybe_raise_if_invalid(ast_value) do
-            !type.kind.non_null?
-          end
-        elsif type.kind.non_null?
-          maybe_raise_if_invalid(ast_value) do
-            (!ast_value.nil?)
-          end && validate(ast_value, type.of_type)
-        elsif type.kind.list?
-          item_type = type.of_type
-          ensure_array(ast_value).all? { |val| validate(val, item_type) }
-        elsif ast_value.is_a?(GraphQL::Language::Nodes::VariableIdentifier)
-          true
-        elsif type.kind.scalar? && constant_scalar?(ast_value)
-          maybe_raise_if_invalid(ast_value) do
-            type.valid_input?(ast_value, @context)
-          end
-        elsif type.kind.enum?
-          maybe_raise_if_invalid(ast_value) do
-            if ast_value.is_a?(GraphQL::Language::Nodes::Enum)
-              type.valid_input?(ast_value.name, @context)
-            else
-              # if our ast_value isn't an Enum it's going to be invalid so return false
-              false
-            end
-          end
-        elsif type.kind.input_object? && ast_value.is_a?(GraphQL::Language::Nodes::InputObject)
-          maybe_raise_if_invalid(ast_value) do
-            required_input_fields_are_present(type, ast_value) && present_input_field_values_are_valid(type, ast_value)
-          end
-        else
-          maybe_raise_if_invalid(ast_value) do
-            false
-          end
+        catch(:invalid) do
+          recursively_validate(ast_value, type)
         end
       end
 
       private
 
+      def recursively_validate(ast_value, type)
+        if type.nil?
+          # this means we're an undefined argument, see #present_input_field_values_are_valid
+          maybe_raise_if_invalid(ast_value) do
+            @invalid_response
+          end
+        elsif ast_value.is_a?(GraphQL::Language::Nodes::NullValue)
+          maybe_raise_if_invalid(ast_value) do
+            type.kind.non_null? ? @invalid_response : @valid_response
+          end
+        elsif type.kind.non_null?
+          maybe_raise_if_invalid(ast_value) do
+            ast_value.nil? ?
+              @invalid_response :
+              recursively_validate(ast_value, type.of_type)
+          end
+        elsif type.kind.list?
+          item_type = type.of_type
+          results = ensure_array(ast_value).map { |val| recursively_validate(val, item_type) }
+          merge_results(results)
+        elsif ast_value.is_a?(GraphQL::Language::Nodes::VariableIdentifier)
+          @valid_response
+        elsif type.kind.scalar? && constant_scalar?(ast_value)
+          maybe_raise_if_invalid(ast_value) do
+            type.validate_input(ast_value, @context)
+          end
+        elsif type.kind.enum?
+          maybe_raise_if_invalid(ast_value) do
+            if ast_value.is_a?(GraphQL::Language::Nodes::Enum)
+              type.validate_input(ast_value.name, @context)
+            else
+              # if our ast_value isn't an Enum it's going to be invalid so return false
+              @invalid_response
+            end
+          end
+        elsif type.kind.input_object? && ast_value.is_a?(GraphQL::Language::Nodes::InputObject)
+          maybe_raise_if_invalid(ast_value) do
+            merge_results([
+              required_input_fields_are_present(type, ast_value),
+              present_input_field_values_are_valid(type, ast_value)
+            ])
+          end
+        else
+          maybe_raise_if_invalid(ast_value) do
+            @invalid_response
+          end
+        end
+      end
+
+      # When `error_bubbling` is false, we want to bail on the first failure that we find.
+      # Use `throw` to escape the current call stack, returning the invalid response.
       def maybe_raise_if_invalid(ast_value)
         ret = yield
-        if !@context.schema.error_bubbling && !ret
-          e = LiteralValidationError.new
-          e.ast_value = ast_value
-          raise e
+        if !@context.schema.error_bubbling && !ret.valid?
+          throw(:invalid, ret)
         else
           ret
         end
@@ -87,27 +101,38 @@ module GraphQL
         present_field_names = ast_node.arguments.map(&:name)
         missing_required_field_names = required_field_names - present_field_names
         if @context.schema.error_bubbling
-          missing_required_field_names.empty?
+          missing_required_field_names.empty? ? @valid_response : @invalid_response
         else
-          missing_required_field_names.all? do |name|
-            validate(GraphQL::Language::Nodes::NullValue.new(name: name), @warden.arguments(type).find { |f| f.name == name }.type )
+          results = missing_required_field_names.map do |name|
+            arg_type = @warden.arguments(type).find { |f| f.name == name }.type
+            recursively_validate(GraphQL::Language::Nodes::NullValue.new(name: name), arg_type)
           end
+          merge_results(results)
         end
       end
 
       def present_input_field_values_are_valid(type, ast_node)
         field_map = @warden.arguments(type).reduce({}) { |m, f| m[f.name] = f; m}
-        ast_node.arguments.all? do |value|
+        results = ast_node.arguments.map do |value|
           field = field_map[value.name]
           # we want to call validate on an argument even if it's an invalid one
           # so that our raise exception is on it instead of the entire InputObject
           type = field && field.type
-          validate(value.value, type)
+          recursively_validate(value.value, type)
         end
+        merge_results(results)
       end
 
       def ensure_array(value)
         value.is_a?(Array) ? value : [value]
+      end
+
+      def merge_results(results_list)
+        merged_result = Query::InputValidationResult.new
+        results_list.each do |inner_result|
+          merged_result.merge_result!([], inner_result)
+        end
+        merged_result
       end
     end
   end