+++ spec/support/star_wars/schema.rb
@@ -48,7 +48,11 @@ module StarWars
     field :total_count, Integer, null: true
 
     def total_count
-      object.nodes.count
+      if TESTING_INTERPRETER
+        object.items.count
+      else
+        object.nodes.count
+      end
     end
   end
 
@@ -158,8 +162,8 @@ module StarWars
     field :shipsWithMaxPageSize, "Ships with max page size", max_page_size: 2, resolver: ShipsWithMaxPageSize
 
     field :bases, BasesConnectionWithTotalCountType, null: true, connection: true do
-      argument :nameIncludes, String, required: false
-      argument :complexOrder, Boolean, required: false
+      argument :name_includes, String, required: false
+      argument :complex_order, Boolean, required: false
     end
 
     def bases(name_includes: nil, complex_order: nil)
@@ -201,7 +205,7 @@ module StarWars
     field :basesWithoutNodes, BaseConnectionWithoutNodes, null: true, resolver_method: :all_bases_array
 
     field :basesAsSequelDataset, BasesConnectionWithTotalCountType, null: true, connection: true, max_page_size: 1000 do
-      argument :nameIncludes, String, required: false
+      argument :name_includes, String, required: false
     end
 
     def bases_as_sequel_dataset(name_includes: nil)
@@ -431,6 +435,8 @@ module StarWars
     if TESTING_INTERPRETER
       use GraphQL::Execution::Interpreter
       use GraphQL::Analysis::AST
+      use GraphQL::Pagination::Connections
+      connections.add(LazyNodesWrapper, LazyNodesRelationConnection)
     end
 
     def self.resolve_type(type, object, ctx)